# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "DeskDibs API",
    "description": "Office hot-desk seat booking. The floor map at GET /api/seatmap is the endpoint the rest of the product hangs off; every mutation broadcasts a SeatStatusChanged message over STOMP to /topic/seatmap/{date} once its transaction commits.",
    "contact": {
      "name": "DeskDibs"
    },
    "version": "v1"
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Generated server url"
    }
  ],
  "security": [
    {
      "bearerAuth": []
    }
  ],
...
  "paths": {
...
    "/api/admin/users/{id}/active": {
      "patch": {
        "tags": [
          "Administration"
        ],
        "summary": "Activate or deactivate an account",
        "description": "Deactivating refuses the account at login and hands back every desk it still holds from today onward. The response names each booking that was released, so the consequence is visible rather than silent. Reactivating restores access but does not reclaim desks.",
        "operationId": "setUserActive",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateUserActiveRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "What changed, including the bookings released.",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/UserActivationReport"
                }
              }
            }
          },
          "403": {
            "description": "The caller is not an administrator.",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/UserActivationReport"
                }
              }
            }
          },
          "404": {
            "description": "No user with that id.",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/UserActivationReport"
                }
              }
            }
          },
          "409": {
            "description": "An administrator may not deactivate their own account.",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/UserActivationReport"
                }
              }
            }
          }
        }
      }
    },
    "/api/admin/seats/{id}/status": {
      "patch": {
        "tags": [
          "Administration"
        ],
        "summary": "Take a desk out of the pool, or put it back",
        "description": "DISABLED or BROKEN withdraws the desk and releases every booking on it from today onward; the affected people see the reason on their own bookings page. ACTIVE returns it to the pool. The seat row is never deleted — booking history survives a floor-plan change.",
        "operationId": "setSeatStatus",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateSeatStatusRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "What changed, including the bookings released.",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/SeatStatusChangeReport"
                }
              }
            }
          },
          "403": {
            "description": "The caller is not an administrator.",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/SeatStatusChangeReport"
                }
              }
            }
          },
          "404": {
            "description": "No seat with that id.",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/SeatStatusChangeReport"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/admin/users": {
      "get": {
        "tags": [
          "Administration"
        ],
        "summary": "Everyone on the system",
        "description": "Name, email, role and whether the account is live. Never returns credentials.",
        "operationId": "users",
        "responses": {
          "200": {
            "description": "The people list.",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AdminUserView"
                  }
                }
              }
            }
          },
          "403": {
            "description": "The caller is not an administrator.",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AdminUserView"
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "UpdateUserActiveRequest": {
        "type": "object",
        "properties": {
          "active": {
            "type": "boolean"
          }
        },
        "required": [
          "active"
        ]
      },
      "ReleasedBookingView": {
        "type": "object",
        "properties": {
          "bookingId": {
            "type": "integer",
            "format": "int64"
          },
          "seatId": {
            "type": "integer",
            "format": "int64"
          },
          "seatLabel": {
            "type": "string"
          },
          "bookingDate": {
            "type": "string",
            "format": "date"
          },
          "userId": {
            "type": "integer",
            "format": "int64"
          },
          "userDisplayName": {
            "type": "string"
          }
        }
      },
      "UserActivationReport": {
        "type": "object",
        "properties": {
          "userId": {
            "type": "integer",
            "format": "int64"
          },
          "displayName": {
            "type": "string"
          },
          "active": {
            "type": "boolean"
          },
          "wasAlreadyInThatState": {
            "type": "boolean"
          },
          "bookingsReleased": {
            "type": "integer",
            "format": "int32"
          },
          "released": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ReleasedBookingView"
            }
          }
        }
      },
      "UpdateSeatStatusRequest": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "enum": [
              "ACTIVE",
              "DISABLED",
              "BROKEN"
            ]
          }
        },
        "required": [
          "status"
        ]
      },
      "SeatStatusChangeReport": {
        "type": "object",
        "properties": {
          "seatId": {
            "type": "integer",
            "format": "int64"
          },
          "seatLabel": {
            "type": "string"
          },
          "previousStatus": {
            "type": "string",
            "enum": [
              "ACTIVE",
              "DISABLED",
              "BROKEN"
            ]
          },
          "status": {
            "type": "string",
            "enum": [
              "ACTIVE",
              "DISABLED",
              "BROKEN"
            ]
          },
          "wasAlreadyInThatState": {
            "type": "boolean"
          },
          "bookingsReleased": {
            "type": "integer",
            "format": "int32"
          },
          "released": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ReleasedBookingView"
            }
          }
        }
      },
...
      "AdminUserView": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "email": {
            "type": "string"
          },
          "displayName": {
            "type": "string"
          },
          "role": {
            "type": "string",
            "enum": [
              "EMPLOYEE",
              "MANAGER",
              "ADMIN"
            ]
          },
          "active": {
            "type": "boolean"
          }
        }
      }
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "description": "Obtain one from POST /api/auth/login (local provider) or the Entra OIDC flow, then send it as \"Authorization: Bearer <token>\".",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}